flask_api/requirements.txt
flask==3.0.2
flask-restful==0.3.10
flask-cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
